// apps/intern/src/app/backend/interfcaes/http-request-options.interface.ts
import { HttpHeaders, HttpParams } from '@angular/common/http';

export interface HttpRequestOptions {
  headers?: HttpHeaders;
  observe?: 'body';
  params?: HttpParams | {
      [param: string]: string | string[];
  };
  reportProgress?: boolean;
  responseType?: 'json';
  withCredentials?: boolean;
}